% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_poilog.R
\name{fit_poilog}
\alias{fit_poilog}
\title{Composite Maximum Likelihood Estimation for Poisson Log-Normal Model}
\usage{
fit_poilog(Y, ncores = 8)
}
\arguments{
\item{Y}{Matrix with proteins counts with dimensions (number of cells) x (number of markers)}

\item{ncores}{Number of CPU cores to use}
}
\value{
A list containing
  \item{beta}{estimated mean parameter}
  \item{Sigma}{estimated covariance parameter}
}
\description{
\code{fit_poilog} uses composite maximum likelihood estimation to fit the mean
and the covariance of a Poisson log-normal mixed model.
}
