% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cytoeffect_poisson.R
\name{plot.cytoeffect_poisson}
\alias{plot.cytoeffect_poisson}
\title{Plot posterior summaries.}
\usage{
plot.cytoeffect_poisson(obj, type = "distribution")
}
\description{
Plot posterior summaries.
}
