% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cytoeffect_poisson_mcle.R
\name{plot.cytoeffect_poisson_mcle}
\alias{plot.cytoeffect_poisson_mcle}
\title{Plot point estimates for MCLE fit}
\usage{
\method{plot}{cytoeffect_poisson_mcle}(x, type = "beta", selection = x$protein_names, ...)
}
\arguments{
\item{x}{Object of class \code{cytoeffect_poisson} computed
using \code{\link{poisson_lognormal}}}

\item{type}{A string with the parameter to plot:
\code{type = "beta"}, \code{type = "sigma"}, or \code{type = "Cor"}}

\item{selection}{A vector of strings with a selection of protein names to plot}

\item{...}{Other parameters}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot point estimates for MCLE fit
}
\examples{
set.seed(1)
df = simulate_data(n_cells = 10)
str(df)
fit = poisson_lognormal_mcle(df,
                             protein_names = names(df)[3:ncol(df)],
                             condition = "condition",
                             group = "donor",
                             ncores = 1)
plot(fit, type = "beta")

}
