% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Posterior multivariate pairs plot for Poisson Log-Normal Mixed Model}
\usage{
plot_pairs(obj, marker1, marker2, marker3)
}
\arguments{
\item{obj}{Object of class \code{cytoeffect_poisson} computed
using \code{\link{poisson_lognormal}}}

\item{marker1}{Name of first marker}

\item{marker2}{Name of second marker}

\item{marker3}{Name of third marker}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Posterior multivariate pairs plot for Poisson Log-Normal Mixed Model
}
\examples{
# fit = cytoeffect::poisson_lognormal(...)
# cytoeffect::plot_pairs(fit,
                         marker1 = "pSTAT1",
                         marker2 = "pSTAT3",
                         marker3 = "pSTAT5")
}
