% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{poisson}
\alias{poisson}
\title{HMC sampling for Poisson Log-Normal Model.}
\usage{
poisson(df_samples_subset, protein_names, condition, group, iter = 325,
  warmup = 200, num_chains = 4, eta = 1)
}
\description{
HMC sampling for Poisson Log-Normal Model.
}
