% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_lognormal.R
\name{poisson_lognormal}
\alias{poisson_lognormal}
\title{HMC sampling for Poisson Log-Normal Model.}
\usage{
poisson_lognormal(df_samples_subset, protein_names, condition, group,
  iter = 325, warmup = 200, num_chains = 4)
}
\description{
HMC sampling for Poisson Log-Normal Model.
}
