% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_lognormal_mcle.R
\name{poisson_lognormal_mcle}
\alias{poisson_lognormal_mcle}
\title{HMC Sampling for Poisson Log-Normal Mixed Model}
\usage{
poisson_lognormal_mcle(
  df_samples_subset,
  protein_names,
  condition,
  group,
  ncores = 4,
  slurm_settings = "slurm_batchtools.tmpl"
)
}
\arguments{
\item{df_samples_subset}{Data frame or tibble with proteins counts,
cell condition, and group information}

\item{protein_names}{A vector of column names of protein to use in the analysis}

\item{condition}{The column name of the condition variable}

\item{group}{The column name of the group variable}

\item{ncores}{Number of CPU cores}

\item{slurm_settings}{Path to slurm cluster template for \code{\link[batchtools]{batchtools}}}
}
\value{
A list of class \code{cytoeffect_poisson_mcle} containing
  \item{tb_args}{output tibble of model fits}
  \item{protein_names}{input protein names}
  \item{condition}{input condition variable}
  \item{group}{input group names}
  \item{df_samples_subset}{input df_samples_subset table}
}
\description{
\code{poisson_lognormal} uses maximum composite likelihood estimation to fit
Poisson log-normal models to each sample.
}
