% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Generate dataset for vignettes and simulation studies}
\usage{
simulate_data(
  n_markers = 5,
  n_true = 1,
  n_samples = 8,
  paired = TRUE,
  n_cells = 100,
  pi = 0.05,
  beta_treatment = log(20),
  beta_control = log(5),
  rho_b = 0.5,
  rho_u = 0.5
)
}
\arguments{
\item{n_markers}{number of markers}

\item{n_true}{number of differential markers}

\item{n_samples}{donors per group}

\item{paired}{paired experimental design}

\item{n_cells}{per sample}

\item{pi}{zero inflation probability}

\item{beta_treatment}{log of mean in treatment}

\item{beta_control}{log of mean in control}

\item{rho_b}{cell level correlation}

\item{rho_u}{donor level correlation}
}
\value{
\code{\link[tibble]{tibble}} data frame
}
\description{
Generate dataset for vignettes and simulation studies
}
\examples{
set.seed(1)
df = simulate_data(n_cells = 10)
str(df)
df

}
