% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmm.R
\name{glmm}
\alias{glmm}
\title{HMC Sampling for Logistic Linear Mixed Model}
\usage{
glmm(
  df_samples_subset,
  protein_names,
  condition,
  group,
  iter = 325,
  warmup = 200,
  num_chains = 4,
  eta = 1
)
}
\arguments{
\item{df_samples_subset}{Data frame or tibble with proteins counts,
cell condition, and group information}

\item{protein_names}{A vector of column names of protein to use in the analysis}

\item{condition}{The column name of the condition variable}

\item{group}{The column name of the group variable}

\item{iter}{Number of iteration per chain for the HMC sampler}

\item{warmup}{Number of warm up steps per chain for the HMC sampler}

\item{num_chains}{Number of HMC chains to run in parallel}

\item{eta}{Hyperparametr for LKJ prior}
}
\value{
A list of class \emph{cytoeffect} containing
  \item{fit_mcmc}{\code{\link[rstan]{rstan}} object}
  \item{protein_names}{input protein names}
  \item{conditions}{input condition variable}
  \item{X}{prediction design matrix}
}
\description{
\code{glmm} uses Hamiltonion Monte Carlo to sample form an extend logistic
linear mixed model. We predict the experimental condition from protein
marker expressions.
}
