% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cytoeffect_poisson.R
\name{plot.cytoeffect_poisson}
\alias{plot.cytoeffect_poisson}
\title{Plot posterior summaries for Poisson log-normal mixed model}
\usage{
\method{plot}{cytoeffect_poisson}(x, type = "beta", selection = x$protein_names, ...)
}
\arguments{
\item{x}{Object of class \code{cytoeffect_poisson} computed
using \code{\link{poisson_lognormal}}}

\item{type}{A string with the parameter to plot:
\code{type = "theta"}, \code{type = "beta"}, \code{type = "sigma"}, or \code{type = "Cor"}}

\item{selection}{A vector of strings with a selection of protein names to plot}

\item{...}{Other parameters}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot posterior summaries for Poisson log-normal mixed model
}
\examples{
set.seed(1)
df = simulate_data(n_cells = 10)
str(df)
fit = poisson_lognormal(df,
                        protein_names = names(df)[3:ncol(df)],
                        condition = "condition",
                        group = "donor",
                        r_donor = 2,
                        warmup = 200, iter = 325, adapt_delta = 0.95,
                        num_chains = 1)
plot(fit, type = "beta")

}
