% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distatis.R
\name{plot_distatis}
\alias{plot_distatis}
\title{DiSTATIS Plot of Posterior or Bootstrap Samples}
\usage{
plot_distatis(
  obj,
  ncores = 1,
  show_donors = TRUE,
  show_markers = TRUE,
  repel = TRUE
)
}
\arguments{
\item{obj}{Object of class \code{cytoeffect_poisson} or \code{cytoeffect_poisson_mcle}
computed by \code{\link{poisson_lognormal}} or \code{\link{poisson_lognormal_mcle}}}

\item{ncores}{Number of cores}

\item{show_donors}{Include donor random effect}

\item{show_markers}{Include markers}

\item{repel}{Repel marker names}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
DiSTATIS Plot of Posterior or Bootstrap Samples
}
\examples{
# fit = cytoeffect::poisson_lognormal(...)
# cytoeffect::plot_distatis(fit)
}
