% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distatis.R
\name{plot_distatis}
\alias{plot_distatis}
\title{Plot Multivariate Posterior Summaries for Poisson Log-Normal Mixed Model}
\usage{
plot_distatis(obj, ncores = parallel::detectCores(),
  show_donors = TRUE, show_markers = TRUE, repel = TRUE)
}
\arguments{
\item{obj}{Object of class \code{cytoeffect_poisson} computed
using \code{\link{poisson_lognormal}}}

\item{ncores}{Number of cores}

\item{show_donors}{Include donor random effect}

\item{show_markers}{Include markers}

\item{repel}{Repel marker names}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot Multivariate Posterior Summaries for Poisson Log-Normal Mixed Model
}
\examples{
# fit = cytoeffect::poisson_lognormal(...)
# cytoeffect::plot_distatis(fit)
}
