% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot Multivariate Posterior Summaries for Poisson Log-Normal Mixed Model}
\usage{
plot_mds(obj, asp = TRUE, ncores = parallel::detectCores(),
  nsubsample = 100, cor_scaling_factor = 1)
}
\arguments{
\item{obj}{Object of class \code{cytoeffect_poisson} computed
using \code{\link{poisson_lognormal}}}

\item{asp}{Set \code{asp = FALSE} to avoid scaling aspect ratio by eigenvalues}

\item{ncores}{Number of cores}

\item{nsubsample}{Subsample size for posterior draws}

\item{cor_scaling_factor}{Scaling factor for correlation arrows}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot Multivariate Posterior Summaries for Poisson Log-Normal Mixed Model
}
\examples{
# fit = cytoeffect::poisson_lognormal(...)
# cytoeffect::plot_mds(fit, asp = FALSE)
}
