% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pairs.R
\name{plot_pairs}
\alias{plot_pairs}
\title{Posterior multivariate pairs plot for Poisson log-normal mixed model}
\usage{
plot_pairs(obj, marker1, marker2, marker3)
}
\arguments{
\item{obj}{Object of class \code{cytoeffect_poisson} computed
using \code{\link{poisson_lognormal}}}

\item{marker1}{Name of first marker}

\item{marker2}{Name of second marker}

\item{marker3}{Name of third marker}
}
\value{
\code{\link[ggplot2]{ggplot2}} object
}
\description{
Posterior multivariate pairs plot for Poisson log-normal mixed model
}
\examples{
set.seed(1)
df = simulate_data(n_cells = 10)
str(df)
fit = poisson_lognormal(df,
                        protein_names = names(df)[3:ncol(df)],
                        condition = "condition",
                        group = "donor",
                        r_donor = 2,
                        warmup = 200, iter = 325, adapt_delta = 0.95,
                        num_chains = 1)
plot_pairs(fit, "m01", "m02", "m03")

}
