% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson_lognormal.R
\name{poisson_lognormal}
\alias{poisson_lognormal}
\title{HMC sampling for Poisson log-normal mixed model}
\usage{
poisson_lognormal(
  df_samples_subset,
  protein_names,
  condition,
  group,
  r_donor,
  eta = 1,
  iter = 325,
  warmup = 200,
  num_chains = 1,
  adapt_delta = 0.8,
  seed = 1
)
}
\arguments{
\item{df_samples_subset}{Data frame or tibble with proteins counts,
cell condition, and group information}

\item{protein_names}{A vector of column names of protein to use in the analysis}

\item{condition}{The column name of the condition variable}

\item{group}{The column name of the group variable}

\item{r_donor}{Rank of the donor random effect covariance matrix}

\item{eta}{Hyperparameter for LKJ prior}

\item{iter}{Number of iteration per chain for the HMC sampler}

\item{warmup}{Number of warm up steps per chain for the HMC sampler}

\item{num_chains}{Number of HMC chains to run in parallel}

\item{adapt_delta}{Parameter to control step size of numerical solver}

\item{seed}{Set seed for HMC sampler
(higher value means smaller step size, max is 1)}
}
\value{
A list of class \code{cytoeffect_poisson} containing
  \item{fit_mcmc}{\code{\link[rstan]{rstan}} object}
  \item{protein_names}{input protein names}
  \item{condition}{input condition variable}
  \item{group}{input group names}
  \item{df_samples_subset}{input df_samples_subset table}
}
\description{
\code{poisson_lognormal} uses Hamiltonion Monte Carlo to sample form an extend
Poisson log-normal mixed model. Each cell and protein marker has its own rate
parameter following a linear model.
}
\examples{
set.seed(1)
df = simulate_data(n_cells = 10)
str(df)
fit = poisson_lognormal(df,
                        protein_names = names(df)[3:ncol(df)],
                        condition = "condition",
                        group = "donor",
                        r_donor = 2,
                        warmup = 200, iter = 325, adapt_delta = 0.95,
                        num_chains = 1)

}
