% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_log_lambda.R
\name{posterior_predictive_log_lambda}
\alias{posterior_predictive_log_lambda}
\title{Posterior predictive of Poisson log-normal mixed model}
\usage{
posterior_predictive_log_lambda(obj, k = 1, show_donors = TRUE)
}
\arguments{
\item{obj}{Object of class \code{cytoeffect_poisson} computed
using \code{\link{poisson_lognormal}}}

\item{k}{Draw from HMC chain}

\item{show_donors}{Include donor random effect}
}
\value{
\code{\link[tibble]{tibble}} object
}
\description{
Posterior predictive of Poisson log-normal mixed model
}
\examples{
set.seed(1)
df = simulate_data(n_cells = 10)
str(df)
fit = poisson_lognormal(df,
                        protein_names = names(df)[3:ncol(df)],
                        condition = "condition",
                        group = "donor",
                        r_donor = 2,
                        warmup = 200, iter = 325,
                        num_chains = 1)
posterior_predictive_log_lambda(fit, k = 1)

}
